+++ src/datasync/messages/registermessage_p.h
@@ -21,12 +21,12 @@ class Q_DATASYNC_EXPORT RegisterMessage
 	Q_PROPERTY(QByteArray cryptAlgorithm MEMBER cryptAlgorithm)
 	Q_PROPERTY(QByteArray cryptKey MEMBER cryptKey)
 	Q_PROPERTY(QString deviceName MEMBER deviceName)
-	Q_PROPERTY(quint32 nonce MEMBER nonce)
+	Q_PROPERTY(quint64 nonce MEMBER nonce)
 
 public:
 	RegisterMessage();
 	RegisterMessage(const QString &deviceName,
-					quint32 nonce,
+					quint64 nonce,
 					const QSharedPointer<CryptoPP::X509PublicKey> &signKey,
 					const QSharedPointer<CryptoPP::X509PublicKey> &cryptKey,
 					AsymmetricCrypto *crypto);
@@ -36,7 +36,7 @@ public:
 	QByteArray cryptAlgorithm;
 	QByteArray cryptKey;
 	QString deviceName;
-	quint32 nonce;
+	quint64 nonce;
 
 	AsymmetricCrypto *createCrypto(QObject *parent = nullptr);
 	QSharedPointer<CryptoPP::X509PublicKey> getSignKey(CryptoPP::RandomNumberGenerator &rng, AsymmetricCrypto *crypto);
